+++ src/encoded/tests/data/cypress-test-inserts/creation-scripts/get_keys.py
@@ -1,0 +1,12 @@
+"""
+Grab CGAP user keys from ~/.cgap-keys.json.
+"""
+
+import json
+from pathlib import Path
+
+
+KEYS_FILE = Path.home().joinpath(".cgap-keys.json")
+
+with KEYS_FILE.open() as file_handle:
+    keys = json.load(file_handle)