+++ CHANGELOG.rst
@@ -6,6 +6,15 @@ cgap-portal
 Change Log
 ----------
 
+11.1.1
+======
+
+`PR 633: Vs cypress test inserts  <https://github.com/dbmi-bgm/cgap-portal/pull/633>`_
+
+* Adds set of inserts for automated Cypress integration tests for UI
+* Includes folder of Python scripts used to create these inserts
+
+
 11.0.1
 ======